+++ tests/AssetlibraryDatabaseTransactions.php
@@ -3,26 +3,25 @@
 namespace Thinktomorrow\AssetLibrary\Tests;
 
 use Illuminate\Filesystem\Filesystem;
+use Illuminate\Database\Schema\Blueprint;
+use Thinktomorrow\AssetLibrary\Tests\stubs\Article;
 
-trait DatabaseTransactions
+trait AssetlibraryDatabaseTransactions
 {
-    protected $connectionsToTransact    = ['testing'];
+    protected $connectionsToTransact = ['testing'];
     protected static $migrationsHaveRun = false;
 
     protected $testDatabasePath = __DIR__.'/../database/testing.sqlite';
 
     protected function setUpDatabase()
     {
-        if (! file_exists($this->testDatabasePath)) {
-            touch($this->testDatabasePath);
-        }
-
         if (! self::$migrationsHaveRun) {
             $this->removeTestDatabase();
 
             touch($this->testDatabasePath);
 
             $migrations = app(Filesystem::class)->allFiles(realpath(__DIR__.'/../database/migrations'));
+
 
             foreach ($migrations as $migration) {
                 include_once $migration->getPathName();
@@ -35,6 +34,8 @@ trait DatabaseTransactions
         }
 
         $this->beginDatabaseTransaction();
+        Article::migrate();
+        Article::create();
     }
 
     protected function removeTestDatabase()
@@ -88,10 +89,10 @@ trait DatabaseTransactions
      */
     private function guessClassNameFromFile($file)
     {
-        $fp    = fopen($file, 'r');
+        $fp = fopen($file, 'r');
         $class = $namespace = $buffer = '';
-        $i     = 0;
-        while (! $class) {
+        $i = 0;
+        while (!$class) {
             if (feof($fp)) {
                 break;
             }
@@ -103,9 +104,9 @@ trait DatabaseTransactions
                 continue;
             }
 
-            for (; $i < count($tokens); $i++) {
+            for (;$i<count($tokens);$i++) {
                 if ($tokens[$i][0] === T_NAMESPACE) {
-                    for ($j=$i + 1; $j < count($tokens); $j++) {
+                    for ($j=$i+1;$j<count($tokens); $j++) {
                         if ($tokens[$j][0] === T_STRING) {
                             $namespace .= '\\'.$tokens[$j][1];
                         } elseif ($tokens[$j] === '{' || $tokens[$j] === ';') {
@@ -115,17 +116,15 @@ trait DatabaseTransactions
                 }
 
                 if ($tokens[$i][0] === T_CLASS) {
-                    for ($j=$i + 1; $j < count($tokens); $j++) {
+                    for ($j=$i+1;$j<count($tokens);$j++) {
                         if ($tokens[$j] === '{') {
-                            $class = $tokens[$i + 2][1];
+                            $class = $tokens[$i+2][1];
                             break;
                         }
                     }
                 }
             }
         }
-
-        fclose($fp);
 
         return $class;
     }